pyproject.toml
[project]
name = "protostocks"
version = "0.1.0"
description = "ProtoStocks project managed with uv and FastAPI."
authors = ["Abhishek Handibag <abhishek.handibag12@gmail.com>"]
requires-python = ">=3.8"
dependencies = [
    "fastapi[standard]>=0.121.0",
    "python-dotenv>=1.0.1",
    "python-jose>=3.4.0",
]

[tool.uv]
# uv-specific configuration can go here
